Added Preliminary Sample Management Domains
Added preliminary Sample and SampleRelationship domain to maintain basic sample information and the relationship between two samples.  This relationship employs a many-to-many relationship, because of the infrequent cases where aliquots, generally of the same Material, are combined into a single tube to run an assay or to create a 'pooled' control.

ExternalIdentifier domain was added to manage samples received from other organizations.

TubeType domain was added to give basic information about the tube holding the Sample.  This may be used later to determine whether a tube can be placed into a storage container and/or as a selection criteria for assays.